+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to the Gainium Backtester library will be documented in this
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.4] - 2026-06-11
+
+### Fixed
+- DIV indicator logic
+
+## [1.6.3] - 2026-06-10
+
+### Changed
+- DCA/combo per-bar deal processing performance.
+
+## [1.6.2] - 2026-06-06
+
+### Changed
+- Expose order origin. 
+
 ## [1.6.1] - 2026-04-06
 
 ### Changed